CB-17746 Make environment's isUpgradeCcmAvailable internal endpoint available from UI
It was annotated with @InternalOnly which makes the call from UI impossible.
